pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seed-monitoring"
version = "1.0.0"
description = "Plugin-based monitoring and alerting system"
authors = [{name = "Your Name", email = "your.email@domain.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pika==1.3.2",
    "PyYAML==6.0.2", 
    "requests==2.32.3",
    "python-dotenv==1.0.1",
    "pymongo==4.8.0",
    "psycopg2-binary==2.9.9"
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0", 
    "mypy>=1.0.0",
    "flake8>=6.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0"
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true